+++ cmd/psqldef/psqldef_test.go
@@ -153,6 +153,21 @@ func TestPsqldefCreateIndex(t *testing.T) {
 	assertApplyOutput(t, createTable, nothingModified)
 }
 
+func TestPsqldefColumnLiteral(t *testing.T) {
+	resetTestDatabase()
+
+	createTable := stripHeredoc(`
+		CREATE TABLE users (
+		  "id" bigint NOT NULL,
+		  "name" text,
+		  "age" integer
+		);
+		`,
+	)
+	assertApplyOutput(t, createTable, applyPrefix+createTable)
+	assertApplyOutput(t, createTable, nothingModified)
+}
+
 func TestPsqldefDataTypes(t *testing.T) {
 	resetTestDatabase()