+++ sqlparser/token_test.go
@@ -57,7 +57,7 @@ func TestLiteralID(t *testing.T) {
 	}}
 
 	for _, tcase := range testcases {
-		tkn := NewStringTokenizer(tcase.in)
+		tkn := NewStringTokenizer(tcase.in, ParserModeMysql)
 		id, out := tkn.Scan()
 		if tcase.id != id || string(out) != tcase.out {
 			t.Errorf("Scan(%s): %d, %s, want %d, %s", tcase.in, id, out, tcase.id, tcase.out)
@@ -130,7 +130,7 @@ func TestString(t *testing.T) {
 	}}
 
 	for _, tcase := range testcases {
-		id, got := NewStringTokenizer(tcase.in).Scan()
+		id, got := NewStringTokenizer(tcase.in, ParserModeMysql).Scan()
 		if tcase.id != id || string(got) != tcase.want {
 			t.Errorf("Scan(%q) = (%s, %q), want (%s, %q)", tcase.in, tokenName(id), got, tokenName(tcase.id), tcase.want)
 		}