+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## unreleased
 
+- Parse PostgreSQL's `"column"` literal properly
 - Dump primary key with `--export` on PostgreSQL
 
 ## v0.3.2